fix rubocop conflicts manual part 3
