+++ app/views/layouts/_footer.html.haml
@@ -1,5 +1,5 @@
 %footer
-  - if current_user
+  - if @current_user.present?
     .vertical-space-xl
       .grid-x.text-center.text-tiny
         .cell.text-muted
