+++ app/views/layouts/application.html.haml
@@ -13,7 +13,7 @@
       .small-12.medium-3.large-2.cell.food-menu-panel
         .title
           food.schi.li
-        - if current_user
+        - if @current_user
           = render 'layouts/navigation'
 
       .small-12.medium-9.large-10.cell
